Display chord-lyric chunks
This change properly parses chunks that have chords and lyrics into
separate sections (<aside> for chords, and <p> for lyrics). It is also
updated to distinguish between lines that have chords (of type "list"
ie that pass the "check_type" function) from ones that do not, since
the latter can be displayed without any special processing.