Add parse support for multi-chords
"Multi-chords" is the name I'm using for brackets containing more than
one chord, for example "[a:m d g c g c1]". It is much easier to write
these in one set of brackets then to put them each in their own
brackets, but it does mean that the parser needs to recognize them.