Adjust alignment and spacing
The buttons were coming out vertical on mobile. I thought I wanted that
but after looking at it I realize it means that the lower buttons can
disappear and the user may not know to look for them. For that reason
I think it is better to have them aligned horizontally and for the
heading size to change instead to leave more space for the buttons.