Move chord checks to "parse"
Errors in chords can be considered candidates for either parse checks
or display checks, since they are both unparsable and undisplayable. At
first I had them in the display category, but here I have moved them to
the parse category. My logic is that these errors should be caught right
when the files are parsed, and that should not wait until the files are
displayed.