Add functions for parsing notes and chords
Previously the chord parsing was actually handled by the "display.py"
script. It makes more sense to me to have the parsing done separately
from the formatting (display), since this will allow for the detection
or errors in the song before any formatting takes place. These new
functions will replace the original display functions, but are not
quite complete yet. Here they are built and tested but not yet
incorporated.