Incorporate new chord parsing and display
After working out the note and chord parsing, it was time to bring that
into production. With this commit, the processing shifts. Before chords
were parsed during the display process. Now files are parsed entirely
in one go: chords are split from lyrics, and chords are split into
constituent parts (root, quality, etc). The benefit of this is that
all parsing is done at once and all errors can be caught at the same
time. Then, in the display process, the information is simply formatted.