Add section types: intro, interlude
I'm not exactly sure how best to support instrumental parts. The section
"instrumental" seems to be of a different category from the rest, since
it describes the type of section as opposed to its role (ie "verse"),
so I think I would rather transition to "intro" and "interlude", but I
will leave support for "instrumental" for now.